pyproject.toml
[project]
name = "discordrpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "colorama>=0.4.6",
    "discord-py>=2.5.2",
    "pydantic-settings>=2.7.1",
]
